fix(rust): 修复 clippy 警告
- init.rs: 移除不必要的类型转换
- path_validator.rs: 折叠 else if 块，使用 &Path 替代 &PathBuf

+++ src-tauri/src/utils/init.rs
@@ -178,7 +178,7 @@ async fn init_logger() -> Result<()> {
                 current_infix: Some("latest"),
                 format: "%Y-%m-%d_%H-%M-%S",
             },
-            Cleanup::KeepLogFiles(log_max_count as usize),
+            Cleanup::KeepLogFiles(log_max_count),
         )
         .filter(Box::new(NoModuleFilter(filters)));
